add scale_candidates function
After extracting candidates locations, I want to get them scaled to
the position in the real images. The scale_candidates function
serves this use-case. It was possible to use the scale_candidate
function in the extrac_roi function and the get_roi_overlay
function, so it also simplified the code.