use bicubic interpolation for tag finding
As the neural network operates on downscaled images it is difficult
to reconstruct the exact tag location in the original image.
This commit adds a bicubic interpolation of the saliency image. This
should lead to a more exact tag position. Nevertheless there exists
still a bias towards the tags being more in the upper left corner.
I figured out that a magic translation about 5 pixel solves that.